mcp-client-example/mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "groq>=0.29.0",
    "mcp>=1.10.1",
    "python-dotenv>=1.1.1",
]
